+++ test/bedrock/raft_test.exs
@@ -222,7 +222,7 @@ defmodule Bedrock.RaftTest do
 
       # Assuming that peer :c receives the ping and responds with a pong,
       # ensure that it is recorded properly.
-      p = Raft.handle_event(p, {:append_entries_ack, 1, true, {0, 0}}, :c)
+      p = Raft.handle_event(p, {:append_entries_ack, 1, true, {0, 0}, {0, 0}}, :c)
 
       verify!()
 
@@ -308,7 +308,11 @@ defmodule Bedrock.RaftTest do
       # and that it has been elected.
 
       expect(MockInterface, :timer, fn :election -> &mock_timer_cancel/0 end)
-      expect(MockInterface, :send_event, fn :c, {:append_entries_ack, 1, true, ^t0} -> :ok end)
+
+      expect(MockInterface, :send_event, fn :c, {:append_entries_ack, 1, true, ^t0, ^t0} ->
+        :ok
+      end)
+
       expect(MockInterface, :leadership_changed, fn {:c, 1} -> :ok end)
 
       p = Raft.handle_event(p, {:append_entries, 1, t0, [], t0}, :c)
@@ -436,7 +440,11 @@ defmodule Bedrock.RaftTest do
 
       expect(MockInterface, :timer, fn :election -> &mock_timer_cancel/0 end)
       expect(MockInterface, :timer, fn :election -> &mock_timer_cancel/0 end)
-      expect(MockInterface, :send_event, fn :c, {:append_entries_ack, 2, true, ^t0} -> :ok end)
+
+      expect(MockInterface, :send_event, fn :c, {:append_entries_ack, 2, true, ^t0, ^t0} ->
+        :ok
+      end)
+
       expect(MockInterface, :leadership_changed, fn {:c, 2} -> :ok end)
 
       p = p |> Raft.handle_event({:append_entries, 2, t0, [], t0}, :c)
@@ -488,8 +496,8 @@ defmodule Bedrock.RaftTest do
         |> Raft.handle_event(:election, :timer)
         |> Raft.handle_event({:vote, 1}, :c)
         |> Raft.handle_event({:vote, 1}, :b)
-        |> Raft.handle_event({:append_entries_ack, 1, true, t0}, :c)
-        |> Raft.handle_event({:append_entries_ack, 1, true, t0}, :b)
+        |> Raft.handle_event({:append_entries_ack, 1, true, t0, t0}, :c)
+        |> Raft.handle_event({:append_entries_ack, 1, true, t0, t0}, :b)
 
       verify!()
 
@@ -513,7 +521,7 @@ defmodule Bedrock.RaftTest do
       expect(MockInterface, :timer, fn :election -> &mock_timer_cancel/0 end)
       expect(MockInterface, :timer, fn :election -> &mock_timer_cancel/0 end)
 
-      expect(MockInterface, :send_event, fn :c, {:append_entries_ack, 2, false, ^t1} ->
+      expect(MockInterface, :send_event, fn :c, {:append_entries_ack, 2, false, ^t1, {0, 0}} ->
         :ok
       end)
 
@@ -557,7 +565,7 @@ defmodule Bedrock.RaftTest do
       expect(MockInterface, :timer, fn :election -> &mock_timer_cancel/0 end)
       expect(MockInterface, :timer, fn :election -> &mock_timer_cancel/0 end)
 
-      expect(MockInterface, :send_event, fn :c, {:append_entries_ack, 2, false, {1, 1}} ->
+      expect(MockInterface, :send_event, fn :c, {:append_entries_ack, 2, false, {1, 1}, {0, 0}} ->
         :ok
       end)
 
@@ -576,7 +584,11 @@ defmodule Bedrock.RaftTest do
 
       expect(MockInterface, :timer, fn :election -> &mock_timer_cancel/0 end)
       expect(MockInterface, :consensus_reached, fn _, ^t1, :behind -> :ok end)
-      expect(MockInterface, :send_event, fn :c, {:append_entries_ack, 2, true, ^t2} -> :ok end)
+
+      expect(MockInterface, :send_event, fn :c, {:append_entries_ack, 2, true, ^t2, ^t2} ->
+        :ok
+      end)
+
       p = p |> Raft.handle_event({:append_entries, 2, t0, [{t1, :data1}, {t2, :data2}], t1}, :c)
 
       verify!()
@@ -588,7 +600,11 @@ defmodule Bedrock.RaftTest do
 
       expect(MockInterface, :timer, fn :election -> &mock_timer_cancel/0 end)
       expect(MockInterface, :consensus_reached, fn _, ^t2, :latest -> :ok end)
-      expect(MockInterface, :send_event, fn :c, {:append_entries_ack, 2, true, ^t2} -> :ok end)
+
+      expect(MockInterface, :send_event, fn :c, {:append_entries_ack, 2, true, ^t2, ^t2} ->
+        :ok
+      end)
+
       p = p |> Raft.handle_event({:append_entries, 2, t2, [], t2}, :c)
 
       verify!()
@@ -736,8 +752,8 @@ defmodule Bedrock.RaftTest do
 
       p =
         p
-        |> Raft.handle_event({:append_entries_ack, 1, true, t0}, :b)
-        |> Raft.handle_event({:append_entries_ack, 1, true, t0}, :c)
+        |> Raft.handle_event({:append_entries_ack, 1, true, t0, t0}, :b)
+        |> Raft.handle_event({:append_entries_ack, 1, true, t0, t0}, :c)
 
       verify!()
       assert p |> Raft.am_i_the_leader?()
@@ -814,7 +830,7 @@ defmodule Bedrock.RaftTest do
       expect(MockInterface, :send_event, fn :b, {:append_entries, 1, ^t2, [], ^t2} -> :ok end)
       expect(MockInterface, :send_event, fn :c, {:append_entries, 1, ^t2, [], ^t2} -> :ok end)
 
-      p = p |> Raft.handle_event({:append_entries_ack, 1, true, t2}, :b)
+      p = p |> Raft.handle_event({:append_entries_ack, 1, true, t2, t2}, :b)
 
       verify!()
       assert t2 == p |> Raft.log() |> Log.newest_transaction_id()
@@ -823,7 +839,7 @@ defmodule Bedrock.RaftTest do
       # :c now acknowledges the heartbeat saying that it has received up to t2
       # as well, and we make a note of that.
 
-      p = p |> Raft.handle_event({:append_entries_ack, 1, true, t2}, :c)
+      p = p |> Raft.handle_event({:append_entries_ack, 1, true, t2, t2}, :c)
 
       verify!()
       assert t2 == p |> Raft.log() |> Log.newest_transaction_id()
@@ -848,7 +864,11 @@ defmodule Bedrock.RaftTest do
       # with the latest transaction we have.
 
       expect(MockInterface, :timer, fn :election -> &mock_timer_cancel/0 end)
-      expect(MockInterface, :send_event, fn :c, {:append_entries_ack, 2, false, ^t1} -> :ok end)
+
+      expect(MockInterface, :send_event, fn :c, {:append_entries_ack, 2, false, ^t1, {0, 0}} ->
+        :ok
+      end)
+
       expect(MockInterface, :leadership_changed, fn {:c, 2} -> :ok end)
 
       p = p |> Raft.handle_event({:append_entries, 2, t1, [{t2, :data1}], t1}, :c)
@@ -858,7 +878,11 @@ defmodule Bedrock.RaftTest do
       # the leader responds with a new set of transactions, beginning where we
       # left off, and we append them to our log. we also acknowledge the leader.
       expect(MockInterface, :timer, fn :election -> &mock_timer_cancel/0 end)
-      expect(MockInterface, :send_event, fn :c, {:append_entries_ack, 2, true, ^t2} -> :ok end)
+
+      expect(MockInterface, :send_event, fn :c, {:append_entries_ack, 2, true, ^t2, ^t2} ->
+        :ok
+      end)
+
       expect(MockInterface, :consensus_reached, fn _, ^t2, :latest -> :ok end)
 
       p = p |> Raft.handle_event({:append_entries, 2, t0, [{t1, :data1}, {t2, :data2}], t2}, :c)
@@ -1075,7 +1099,7 @@ defmodule Bedrock.RaftTest do
       expect(MockInterface, :timer, fn :election -> &mock_timer_cancel/0 end)
       expect(MockInterface, :leadership_changed, fn {:undecided, 2} -> :ok end)
 
-      p = Raft.handle_event(leader, {:append_entries_ack, 2, true, {1, 1}}, :peer_d)
+      p = Raft.handle_event(leader, {:append_entries_ack, 2, true, {1, 1}, {1, 1}}, :peer_d)
 
       assert %Raft{mode: %Follower{term: 2, leader: :undecided}} = p
       expect(MockInterface, :timer, fn :election -> &mock_timer_cancel/0 end)
@@ -1083,7 +1107,7 @@ defmodule Bedrock.RaftTest do
 
       expect(MockInterface, :send_event, fn :actual_leader,
                                             {:append_entries_ack, 2, true,
-                                             ^initial_transaction_id} ->
+                                             ^initial_transaction_id, ^initial_transaction_id} ->
         :ok
       end)